backend/requirements.txt
# backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
playwright==1.38.0
schedule==1.2.0
asyncio
python-dotenv==1.0.0